Logic improvement
Separted classes from main file
